# Feature/improve performance of portfolio snapshot computation